requirements.txt
pynacl
PyYaml
psutil
python-dotenv
git+https://github.com/Pycord-Development/pycord